+++ com.unity.render-pipelines.universal/Editor/Converter/RenderPipelineConvertersEditor.cs
@@ -78,13 +78,11 @@ namespace UnityEditor.Rendering.Universal
 
         TypeCache.TypeCollection m_ConverterContainers;
 
-#if RENDER_PIPELINE_CONVERTER
-        [MenuItem("RenderPipelineConverter/RenderPipelineConverter")]
-#endif
-
+        [MenuItem("Window/Rendering/Render Pipeline Converter", false, 50)]
         public static void ShowWindow()
         {
             RenderPipelineConvertersEditor wnd = GetWindow<RenderPipelineConvertersEditor>();
+            wnd.titleContent = new GUIContent("Render Pipeline Converter");
             DontSaveToLayout(wnd);
 
             wnd.Show();